+++ src/test/java/org/folio/entlinks/controller/InstanceAuthorityLinkStatisticsIT.java
@@ -2,6 +2,11 @@ package org.folio.entlinks.controller;
 
 import static java.util.Collections.singletonList;
 import static org.folio.entlinks.domain.dto.LinkAction.UPDATE_HEADING;
+import static org.folio.support.MatchUtils.errorCodeMatch;
+import static org.folio.support.MatchUtils.errorMessageMatch;
+import static org.folio.support.MatchUtils.errorTotalMatch;
+import static org.folio.support.MatchUtils.errorTypeMatch;
+import static org.folio.support.MatchUtils.statsMatch;
 import static org.folio.support.TestDataUtils.Link.TAGS;
 import static org.folio.support.TestDataUtils.linksDto;
 import static org.folio.support.TestDataUtils.linksDtoCollection;
@@ -12,7 +17,6 @@ import static org.folio.support.base.TestConstants.authorityStatsEndpoint;
 import static org.folio.support.base.TestConstants.inventoryAuthorityTopic;
 import static org.folio.support.base.TestConstants.linksInstanceEndpoint;
 import static org.folio.support.base.TestConstants.linksStatsInstanceEndpoint;
-import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.empty;
 import static org.hamcrest.Matchers.hasSize;
@@ -29,18 +33,14 @@ import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.time.temporal.ChronoUnit;
-import java.util.Collection;
-import java.util.LinkedHashMap;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Objects;
 import java.util.UUID;
 import java.util.stream.Stream;
 import lombok.SneakyThrows;
 import org.apache.commons.lang3.ThreadUtils;
 import org.folio.entlinks.domain.dto.AuthorityInventoryRecord;
 import org.folio.entlinks.domain.dto.AuthorityInventoryRecordMetadata;
-import org.folio.entlinks.domain.dto.BibStatsDto;
 import org.folio.entlinks.domain.dto.BibStatsDtoCollection;
 import org.folio.entlinks.domain.dto.LinkStatus;
 import org.folio.entlinks.exception.type.ErrorCode;
@@ -50,9 +50,6 @@ import org.folio.support.DatabaseHelper;
 import org.folio.support.TestDataUtils;
 import org.folio.support.base.IntegrationTestBase;
 import org.folio.support.base.TestConstants;
-import org.hamcrest.BaseMatcher;
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.web.servlet.ResultMatcher;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
@@ -319,18 +316,6 @@ class InstanceAuthorityLinkStatisticsIT extends IntegrationTestBase {
     return jsonPath("$.next", is(next));
   }
 
-  private ResultMatcher statsMatch(Matcher<Collection<? extends BibStatsDto>> matcher) {
-    return jsonPath("$.stats", matcher);
-  }
-
-  @SuppressWarnings("unchecked")
-  private ResultMatcher statsMatch(BibStatsDtoCollection stats) {
-    var statsMatchers = stats.getStats().stream()
-      .map(StatsMatcher::statsMatch)
-      .toArray(Matcher[]::new);
-    return jsonPath("$.stats", contains(statsMatchers));
-  }
-
   private void sendInventoryAuthorityEvent(UUID authorityId, String type) {
     var authUpdateEvent = TestDataUtils.authorityEvent(type,
       new AuthorityInventoryRecord().id(authorityId).personalName("new personal name").naturalId("naturalId")
@@ -339,38 +324,5 @@ class InstanceAuthorityLinkStatisticsIT extends IntegrationTestBase {
       new AuthorityInventoryRecord().id(authorityId).personalName("personal name").naturalId("naturalId"));
     sendKafkaMessage(inventoryAuthorityTopic(), authorityId.toString(), authUpdateEvent);
 
-  }
-
-  private static final class StatsMatcher extends BaseMatcher<BibStatsDto> {
-
-    private final BibStatsDto expectedStats;
-
-    private StatsMatcher(BibStatsDto expectedStats) {
-      this.expectedStats = expectedStats;
-    }
-
-    static StatsMatcher statsMatch(BibStatsDto expectedStats) {
-      return new StatsMatcher(expectedStats);
-    }
-
-    @Override
-    @SuppressWarnings("rawtypes")
-    public boolean matches(Object actual) {
-      if (actual instanceof LinkedHashMap actualStats) {
-        return Objects.equals(expectedStats.getInstanceId().toString(), actualStats.get("instanceId"))
-          && Objects.equals(expectedStats.getAuthorityNaturalId(), actualStats.get("authorityNaturalId"))
-          && Objects.equals(expectedStats.getBibRecordTag(), actualStats.get("bibRecordTag"))
-          && Objects.equals(expectedStats.getInstanceTitle(), actualStats.get("instanceTitle"))
-          && expectedStats.getUpdatedAt().isAfter(OffsetDateTime.parse((String) actualStats.get("updatedAt")))
-          && Objects.equals(expectedStats.getErrorCause(), actualStats.get("errorCause"));
-      }
-
-      return false;
-    }
-
-    @Override
-    public void describeTo(Description description) {
-      description.appendValue(expectedStats);
-    }
   }
 }